delay proof-confirm-complete until no incomplete sequent remains
The remaining source of late show-goal commands were show-goal
requests for incomplete sequents, which were still on the way to
Proof General when the proof complete message arrived. Only once
the update arrived in proof-tree all the show-goal requests for
that sequent could be issued.

Count incomplete sequnets and delay the proof-confirm-complete
message until this is zero. The check for this comes after
registering sequents on evars, therefore the show-goal commands
for evar instantiations precede the proof-confirm-complete
message.